refactor: 플레이어 이름 원시값 포장
